🚧 WIP: RoomController alterado, li de salas adicionado. Incompleto!

+++ app/Http/Controllers/Schools/Rooms/RoomsController.php
@@ -14,8 +14,6 @@ use App\Models\School\School;
 class RoomsController extends Controller
 {
     /**
-     *  git config --global user.email "you@example.com"
-  git config --global user.name "Your Name"
      * Store a newly created resource in storage.
      */
     public function store(StoreRoomsRequest $request): mixed
@@ -34,10 +32,10 @@ class RoomsController extends Controller
      */
     public function update(UpdateRoomsRequest $request): mixed
     {
-        $Room = Rooms::where('code', $request->room_code)->first();
-        $Room->update([
-            'name' => $request->name != null ? $request->name : $Room->name,
-            'description' => $request->description != null ? $request->description : $Room->description,
+        $room = Rooms::where('code', $request->room_code)->first();
+        $room->update([
+            'name' => $request->name != null ? $request->name : $room->name,
+            'description' => $request->description != null ? $request->description : $room->description,
         ]);
 
         return $this->response($request, 'manage.rooms', 'Sala atualizada com sucesso.');
@@ -50,9 +48,9 @@ class RoomsController extends Controller
     {
         $room = Rooms::where('code', $request->room_code)->first();
 
-        // if ($room->classes) {
-        //     return $this->response($request, 'manage.rooms', 'Não é possível deletar uma sala que está sendo utilizada.', 'error', 400);
-        // }
+        if ($room->classes->count() > 0) {
+             return $this->response($request, 'manage.rooms', 'Não é possível deletar uma sala que está sendo utilizada.', 'error', 400);
+        }
 
         $room->delete();
         return $this->response($request, 'manage.rooms', 'Sala deletada com sucesso.');
